rm: E2E test is a waste of time
