+++ Rust_part/tests/consts.rs
@@ -1,0 +1,1 @@
+pub static TEST_IMAGE: &'static str = "https://i.imgur.com/j71oECe.png";
